+++ js/views/abstract-view.js
@@ -1,0 +1,29 @@
+import {getElementFromTemplate} from '../utils/util';
+
+export default class AbstractView {
+  constructor() {
+    if (new.target === AbstractView) {
+      throw new Error(`Can't instantiate AbstractView, only concrete one`);
+    }
+  }
+
+  get template() {
+    throw new Error(`Template is required`);
+  }
+
+  get element() {
+    if (this._element) {
+      return this._element;
+    }
+    this._element = this.render();
+    this.bind(this._element);
+    return this._element;
+  }
+
+  render() {
+    return getElementFromTemplate(this.template);
+  }
+
+  bind() {
+  }
+}
